ui package structure refactoring
