+++ app/src/main/java/ua/blackwind/data/cat_facts/CatFactsRepository.kt
@@ -3,7 +3,6 @@ package ua.blackwind.data.cat_facts
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.adapter
 import kotlinx.coroutines.*
-import kotlinx.coroutines.Dispatchers.IO
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.*
 import org.json.JSONArray
