+++ app/src/main/java/ua/blackwind/ui/screens/main/CatFactCard.kt
